+++ packages/web-app-files/tests/unit/mixins/actions/delete.spec.js
@@ -67,6 +67,7 @@ function getWrapper({ deletePermanent = false, invalidLocation = false } = {}) {
     localVue,
     store,
     mocks: {
+      space: { driveType: 'personal', spaceRoles: { viewer: [], editor: [], manager: [] } },
       $router: {
         currentRoute: invalidLocation
           ? createLocationShares('files-shares-via-link')