+++ packages/web-pkg/src/services/eventBus.ts
@@ -36,3 +36,5 @@ export class EventBus {
     )
   }
 }
+
+export const eventBus = new EventBus()