+++ packages/web-runtime/src/layouts/Plain.vue
@@ -3,3 +3,11 @@
     <router-view />
   </div>
 </template>
+
+<script lang="ts">
+import { defineComponent } from '@vue/composition-api'
+
+export default defineComponent({
+  name: 'PlainLayout'
+})
+</script>