+++ packages/web-runtime/src/pages/resolvePublicLink.vue
@@ -71,24 +71,27 @@
   </div>
 </template>
 
-<script type="ts">
+<script lang="ts">
 import { mapGetters } from 'vuex'
 import { SharePermissionBit } from 'web-client/src/helpers/share'
-import { authService } from "../services/auth";
+import { authService } from '../services/auth'
 
 import {
   queryItemAsString,
   useClientService,
   useRouteParam,
   useRouteQuery
-} from "web-pkg/src/composables";
+} from 'web-pkg/src/composables'
 import { useTask } from 'vue-concurrency'
-import { ref, unref, computed, defineComponent } from "@vue/composition-api";
-import { buildPublicSpaceResource, isPublicSpaceResource } from "web-client/src/helpers";
-import { buildWebDavPublicPath } from "files/src/helpers/resources";
-import isEmpty from "lodash-es/isEmpty";
+import { ref, unref, computed, defineComponent } from '@vue/composition-api'
+import {
+  buildPublicSpaceResource,
+  isPublicSpaceResource,
+  PublicSpaceResource
+} from 'web-client/src/helpers'
+import { buildWebDavPublicPath } from 'files/src/helpers/resources'
+import isEmpty from 'lodash-es/isEmpty'
 import { useLoadTokenInfo } from '../composables/tokenInfo'
-
 
 export default defineComponent({
   name: 'ResolvePublicLink',
@@ -96,19 +99,25 @@ export default defineComponent({
     const { webdav } = useClientService()
     const token = useRouteParam('token')
     const password = ref('')
-    const publicLinkSpace = computed(() => buildPublicSpaceResource({
-      id: unref(token),
-      driveAlias: `public/${unref(token)}`,
-      driveType: 'public',
-      webDavPath: buildWebDavPublicPath(unref(token), ''),
-      ...(unref(password) && { publicLinkPassword: unref(password) })
-    }))
+    const publicLinkSpace = computed(() =>
+      buildPublicSpaceResource({
+        id: unref(token),
+        driveAlias: `public/${unref(token)}`,
+        driveType: 'public',
+        webDavPath: buildWebDavPublicPath(unref(token), ''),
+        ...(unref(password) && { publicLinkPassword: unref(password) })
+      })
+    )
     const isPasswordRequiredTask = useTask(function* (signal, ref) {
       if (!isEmpty(ref.tokenInfo)) {
         return ref.tokenInfo.password_protected
       }
       try {
-        yield webdav.getFileInfo({ ...unref(publicLinkSpace), publicLinkPassword: null })
+        let space: PublicSpaceResource = {
+          ...unref(publicLinkSpace),
+          publicLinkPassword: null
+        }
+        yield webdav.getFileInfo(space)
         return false
       } catch (error) {
         if (error.statusCode === 401) {
@@ -120,7 +129,7 @@ export default defineComponent({
     const loadPublicLinkTask = useTask(function* () {
       const resource = yield webdav.getFileInfo(unref(publicLinkSpace))
       if (!isPublicSpaceResource(resource)) {
-        const e = new Error("resolved resource has wrong type")
+        const e: any = new Error('resolved resource has wrong type')
         e.resource = resource
         throw e
       }
@@ -175,7 +184,10 @@ export default defineComponent({
   methods: {
     async resolvePublicLink(passwordRequired) {
       if (this.tokenInfo?.alias_link) {
-        return this.$router.push({ name: 'resolvePrivateLink', params: { fileId: this.tokenInfo.id }})
+        return this.$router.push({
+          name: 'resolvePrivateLink',
+          params: { fileId: this.tokenInfo.id }
+        })
       }
 
       const publicLink = await this.loadPublicLinkTask.perform()
@@ -197,7 +209,10 @@ export default defineComponent({
         return this.$router.push({ name: 'files-public-upload', params: { token: this.token } })
       }
 
-      return this.$router.push({ name: 'files-public-link', params: { driveAliasAndItem: `public/${this.token}` } })
+      return this.$router.push({
+        name: 'files-public-link',
+        params: { driveAliasAndItem: `public/${this.token}` }
+      })
     }
   }
 })