+++ CHANGELOG.md
@@ -26,8 +26,10 @@ Summary
 * Bugfix - Shares loading: [#7506](https://github.com/owncloud/web/issues/7506)
 * Bugfix - Sidebar toggle icon: [#7632](https://github.com/owncloud/web/pull/7632)
 * Bugfix - Try to obtain refresh token before the error case: [#7756](https://github.com/owncloud/web/pull/7756)
+* Bugfix - Hide actions in space trash bins: [#7768](https://github.com/owncloud/web/pull/7768)
 * Bugfix - Spaces on "Shared via link"-page: [#7651](https://github.com/owncloud/web/pull/7651)
 * Bugfix - Spaces reactivity on update: [#7521](https://github.com/owncloud/web/issues/7521)
+* Bugfix - Trash bin sidebar: [#7778](https://github.com/owncloud/web/issues/7778)
 * Bugfix - Upload modify time: [#7630](https://github.com/owncloud/web/pull/7630)
 * Change - Drive aliases in URLs: [#6648](https://github.com/owncloud/web/issues/6648)
 * Enhancement - Edit custom permissions wording: [#7709](https://github.com/owncloud/web/pull/7709)
@@ -146,6 +148,7 @@ Details
 
    https://github.com/owncloud/web/issues/7347
    https://github.com/owncloud/web/pull/7771
+   https://github.com/owncloud/web/pull/7790
 
 * Bugfix - Search bar on small screens: [#7675](https://github.com/owncloud/web/pull/7675)
 
@@ -199,6 +202,13 @@ Details
 
    https://github.com/owncloud/web/pull/7756
 
+* Bugfix - Hide actions in space trash bins: [#7768](https://github.com/owncloud/web/pull/7768)
+
+   Actions in space trash bins are now hidden if the current user has insufficient permissions.
+
+   https://github.com/owncloud/web/issues/7702
+   https://github.com/owncloud/web/pull/7768
+
 * Bugfix - Spaces on "Shared via link"-page: [#7651](https://github.com/owncloud/web/pull/7651)
 
    Spaces on the "Shared via link"-page are now being displayed correctly. Also, the sidebar for
@@ -215,6 +225,14 @@ Details
 
    https://github.com/owncloud/web/issues/7521
    https://github.com/owncloud/web/pull/7546
+
+* Bugfix - Trash bin sidebar: [#7778](https://github.com/owncloud/web/issues/7778)
+
+   We've fixed the sidebar in the trash bin which was throwing errors and not showing the right
+   content.
+
+   https://github.com/owncloud/web/issues/7778
+   https://github.com/owncloud/web/pull/7787
 
 * Bugfix - Upload modify time: [#7630](https://github.com/owncloud/web/pull/7630)
 
@@ -237,6 +255,7 @@ Details
 
    https://github.com/owncloud/web/issues/6648
    https://github.com/owncloud/web/pull/7430
+   https://github.com/owncloud/web/pull/7791
 
 * Enhancement - Edit custom permissions wording: [#7709](https://github.com/owncloud/web/pull/7709)